+++ Misc/NEWS.d/next/C API/2023-07-01-21-23-33.gh-issue-106316.hp2Ijw.rst
@@ -1,0 +1,2 @@
+Remove ``cpython/pytime.h`` header file: it only contained private
+functions. Patch by Victor Stinner.
