+++ PCbuild/_testcapi.vcxproj.filters
@@ -30,9 +30,6 @@
     <ClCompile Include="..\Modules\_testcapi\unicode.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\Modules\_testcapi\pytime.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="..\Modules\_testcapi\datetime.c">
       <Filter>Source Files</Filter>
     </ClCompile>
